pulley: Ungate memory64 feature (#9780)
* pulley: Ungate memory64 feature

This commit is similar to #9779 in that it's removing a proposal from
the "known list of panicking features" for Pulley to allow more tests to
run on Pulley. This then fills out a few miscellaneous instructions to
get a full suite of tests passing in Pulley related to memory64 and
other instructions.

prtest:full

* Add notes about spectre speculation