pulley: Update how tests are expected to fail or pass (#9786)
With enough basic support here and there switch the listing of tests for
Pulley from known-passing to known-failing instead. This trims the list
of unsupported features to just those that are truly unsupported and
then additionally rewrites the list of tests to be an exhasutive list of
unsupported tests instead of a list of supported tests.

This should make it a bit easier to track the progress as now pulley is
"complete" once this list is empty.