pulley: Fill out remaining 32/64-bit integer operations  (#9803)
* pulley: Fill out remaining 32/64-bit integer operations

This required some extra plumbing to shepherd the precise reason why
signed division trapped to Wasmtime which is done through an extra
`TrapKind` side channel now added.

This then additionally fixes the signed remainder interpreter function
to return 0 on `MIN % -1` which is different from what Rust specifies
(which is to return `None` or panic).

cc #9783

* Reduce some code duplication

* Fix pulley tests

* Fix MSRV compat