pulley: Run many existing `*.clif` runtests (#9825)
This commit adds the pulley targets to many of the preexisting `*.clif`
runtests throughout the tree. This covers most of the MVP functionality
of wasm for example and additionally exercises 8 and 16-bit lowerings
for many instructions. A few minor pulley instructions were added and
otherwise new 8/16-bit lowerings use existing instructions. It's
expected that the 8/16-bit lowerings won't be used all that often so
they're not particularly optimal at this time.

Some CLIF tests were omitted such as:

* Most SIMD-using CLIF tests
* Float/int conversion tests using 8 and 16-bit integers
* Tests with `call` instructions as relocations don't work with the JIT
  crate on Pulley
* Tests using 128-bit integers

Support for some of these tests may be enabled in the future, but for
example 8/16-bit integers may not get used much.