cranelift: Restore 16-bit TrapCodes
This allows users of Cranelift to have ~65k unique trap codes
again.  The range of `TrapCode` was reduced recently in order to
be able to fit arbitrary trap codes inside of `MemFlags`.

We keep the size of MemFlags with this change and choose to only
allow a _low subset_ of trap codes to be represented instead.
Explicit traps can still use the full TrapCode range.