Move memories first in `VMContext` (#9847)
* Move memories first in `VMContext`

This commit shuffles the fields of `VMContext` to move memories first
above all other items. This is done to ensure that the offset to
memory-related fields is one of the smaller offsets and ideally fits within an
8-byte offset which helps fit into some special cases on various ISAs.
For example x64 has a smaller instruction encoding for 8-bit offsets
than 32-bit offsets, and I'm planning on doing a similar optimization
for Pulley.

* Review comments