Optimize constant offset loads/stores in translation (#9845)
* Optimize constant offset loads/stores in translation

This commit implements an optimization when lowering Wasm bytecode to
CLIF to skip a bounds check when the offset in memory is statically
known. This comes up in C/C++/Rust when `static` memory is addressed for
example where LLVM emits an `i32.const 0` as the base address and puts
the address of the variable in the `offset` instruction field. This
isn't necessary for 64-bit platforms but when explicit bounds-checks are
required this can help to eliminate a constant-true bounds-check.

* Review comments