pulley: Add macro instructions for function prologue/epilogue
This commit adds two new instructions to Pulley to combine the
operations of setting up a frame, allocating stack, and saving clobbered
registers. This is all combined into a single instruction which is
relatively large but is much smaller than each of these individual
operations exploded out.

This is a size win on `spidermonkey.cwasm` by about 1M and locally in a
small `fib.wat` test this is also a good speedup by reducing the number
of instructions executed.