Test x86_64-apple-darwin on newer CI machines (#9880)
This commit moves testing of x86_64-apple-darwin from `macos-13` macs
to the `macos-14` image which is newer. The hope is that these newer
images, being natively M1 machines, are faster to compile and less
flaky. A number of CI builds recently have failed due to seeming CI
issues with the older machines, so this'll hopefully keep the same
amount of testing while improving robustness.