Refactor release builds, add aarch64-musl
This commit refactors the way release builds are done in CI in terms of
configuration and then additionally adds aarch64-musl release artifacts
as requested in #9875. The refactoring here is done to reduce the number
of locations to understand release builds. Notably the
`binary-compatible-builds` action was removed in favor of direct
environment configuration in conjunction with docker images used to
build. The `aarch64-musl` build itself happens in a container provided
by the `cross` project to ensure that the right toolchain is configured.

Closes #9875

prtest:full