Fix incorrect frame code offset in coredump
Currently, coredump include incorrect code offsets, the generation relies
on `func_offset()` which returns the instruction's offset relative to the
function.

This change uses `module_offset()` to get the absolute code offset and,
using a new method `get_relative_codeoffset`, converts the offset to be
relative to the code section.

Also, the `name` property has been removed because it's misleading when
displayed and not actually useful.