Clarify MSRV intent in contributor docs
Currently we support 3 versions of rustc (current, current-1, and
current-2). This is probably too small a window for someone out there on
the internet so this commit adds a clarification with some rationale for
the current policy as well as an olive branch for "please talk to us if
you want a wider window". Basically I wanted to update this to
explicitly declare intent that we're open to reconsidering this policy
as necessary (without actually committing to any alternative policy
yet).