Move gdbjit C helpers to a separate file (#9939)
* Move gdbjit C helpers to a separate file

This commit splits out the gdbjit-related helpers from `helpers.c` in
Wasmtime to a separate C file built as part of the `wasmtime-jit-debug`
crate. This'll help excise these helpers if gdbjit support is disabled
at compile time and additionally brings them closer to the actual
definition in the `wasmtime-jit-debug` crate.

* Fix linkage issues

* Fix miri tests