Update to the Rust 2024 edition
This is a test to see what breaks in CI when updating to the 2024
edition. The edition won't actually be released until February of this
year and we won't be able to use it until two releases after that. This
is just a test in the meantime to smooth the future transition.