pulley: Fix a panic in using a return pointer
This fixes a panic in the pulley backend introduced in #9874 which comes
up when using a return pointer on call instructions. This would shuffle
the `uses` list of registers to not be sorted like the assertion was
expecting. The fix in this commit is to just go ahead and sort the list
of `uses` to ensure that registers are peeled off in order.