Emit post-optimization CLIF with `--emit-clif`
This commit updates the implementation of the `wasmtime` CLI flag
`--emit-clif` to emit the IR post-optimization rather than
pre-optimization. This is now emitting the IR that's directly fed to the
backend for lowering which reflects any mid-level transformation that's
performed. It should still be possible to recover pre-optimized CLIF
with `-O opt-level=0`.

This then additionally refactors functions to emit CLIF in a more "core"
location so all trampolines get their CLIF emitted as well in case those
are needed for debugging.