pulley: Fix regalloc of return-call-indirect (#10021)
This commit fixes an issue in the Pulley backend for the
`return_call_indirect` implementation. This brings Pulley in line with
other backends to use a fixed register for the indirect call location
which is caller-save instead of possibly using a callee-save register.
If a callee-save-register is used then the location to be jumped to is
clobbered by register restores and won't have the correct location to
jump to.

This additionally required updating the Pulley ABI slightly. Previously
all caller-saved registers were considered argument registers meaning
that there weren't any actual available registers to use for the jump
destination. To handle this I've decreased the number of argument
registers by 1 so there's a single register used for the
return-call-indirect destination available which is also caller-saved.