Disable parallel compilation during fuzzing (#10035)
The intention has always been to disable parallel compilation during
fuzzing and this was previously achieved with a single-thread pool for
Rayon. This still spawns a rayon thread though and can offload work to
it, so this instead explicitly uses `wasmtime::Config` to disable
parallel compilation. This should ensure that `rayon` doesn't get used
at runtime, as desired, and additionally avoids spawning threads or
offloading work onto a thread.