Disable relaxed-simd in differential fuzzing (#10045)
We forgot to do this quite a long time ago but this was always the
intention. Turns out with Pulley now being online the fuzzer quickly
found a difference between Pulley and Cranelift where the native x64
instructions differ from the "deterministic" behavior that Pulley
implements. This difference is expected and allowed though, so don't
fuzz it.