winch: Refactor global address calculation
This commit fixes a fuzz-bug. See the test case for details.

Prior to this commit, imported global addresses were calculated with
register offset addressing, using the scratch register as the base. With
imported globals, the caller must load the address into an allocatable
register which implies that in presence of spills the scratch register
would get clobbered, affecting the previously loaded imported global
address.

This commit fixes the issue by returning an allocatable register, along
with the offset and global type, which is expected to be freed by the
caller after emitting the global load or store.