pulley: Slightly optimize bounds checks
In profiling a module I was noticing that the previous
`xbc32_bound_trap` instruction wasn't being used when I expected.
Investigation revealed that the load of the bound itself was GVN'd and
deduplicated (yay!) but it meant that the load was used in two locations
meaning it didn't pass checks for `sinkable_load`. This commit fixes
this by repurposing `xbc32_bound_trap` for "the bound is in a register"
and renaming the previous instruction to `xbc32_boundne_trap`. This
helps cut down on the number of opcodes in this benchmark and improves
performance slightly.

At the same time this tightens up "sinkable loads" to require native
endianness since that's what the bound of memory is stored as.
Additionally in addition to testing for `a < b` and optimizing that this
also now optimizes `b > a`, the same condition just having the arguments
swapped.